+++ netket/_src/driver/abstract_variational_driver_old.py
@@ -197,7 +197,7 @@ class AbstractVariationalDriver(abc.ABC):
         self._optimizer = optimizer
         if optimizer is not None:
             self._optimizer_state = optimizer.init(self.state.parameters)
-            if config.netket_experimental_sharding:
+            if config.netket_sharding:
                 self._optimizer_state = jax.lax.with_sharding_constraint(
                     self._optimizer_state,
                     NamedSharding(jax.sharding.get_abstract_mesh(), P()),
@@ -433,7 +433,7 @@ def apply_gradient(optimizer_fun, optimizer_state, dp, params):
 
     new_params = optax.apply_updates(params, updates)
 
-    if config.netket_experimental_sharding:
+    if config.netket_sharding:
         sharding = NamedSharding(jax.sharding.get_abstract_mesh(), P())
         new_optimizer_state = jax.lax.with_sharding_constraint(
             new_optimizer_state, sharding