+++ netket/jax/_vjp_chunked.py
@@ -301,7 +301,7 @@ def vjp_chunked(
     ############################################################################
     # sharding
 
-    if config.netket_experimental_sharding and chunk_size is not None:
+    if config.netket_sharding and chunk_size is not None:
         # assume the chunk_argnums are also sharded
         # later we might introduce an extra arg for it
         sharded_argnums = chunk_argnums